trade-agent/requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
pydantic>=2.7.0
python-dotenv>=1.0.0
requests>=2.32.0
pandas>=2.2.0
sqlalchemy>=2.0.0
pyarrow>=15.0.0
protobuf>=5.27.0
google-generativeai>=0.7.0
scikit-learn>=1.5.0
joblib>=1.4.2
lightgbm>=4.2.0

